shared/vertex_registry.py
"""Vertex AI Agent Registry integration for agent discovery and registration"""

import logging
from typing import Dict, List, Optional, Any
from google.cloud import aiplatform
from google.auth import default

logger = logging.getLogger(__name__)


class VertexAIAgentRegistry:
    """Manages agent registration and discovery in Vertex AI Agent Registry"""
    
    def __init__(self, project_id: str, location: str = "us-central1"):
        self.project_id = project_id
        self.location = location
        aiplatform.init(project=project_id, location=location)
        self.credentials, _ = default()
    
    def register_agent(
        self,
        agent_name: str,
        endpoint: str,
        capabilities: List[str],
        metadata: Optional[Dict[str, Any]] = None
    ) -> str:
        """
        Register an agent in Vertex AI Agent Registry
        
        Args:
            agent_name: Name of the agent
            endpoint: HTTPS endpoint URL for the agent
            capabilities: List of capabilities the agent provides
            metadata: Optional metadata about the agent
        
        Returns:
            Agent resource name in Vertex AI
        """
        try:
            # Create agent resource in Vertex AI
            # Note: This is a simplified version - actual implementation would use
            # Vertex AI Agent Builder API or custom metadata service
            
            agent_resource = {
                "name": agent_name,
                "endpoint": endpoint,
                "capabilities": capabilities,
                "metadata": metadata or {},
                "status": "ACTIVE"
            }
            
            # In production, this would create an actual Vertex AI resource
            # For now, we'll log and return a placeholder
            logger.info(f"Registered agent {agent_name} with endpoint {endpoint}")
            logger.info(f"Capabilities: {capabilities}")
            
            # Return resource name (format: projects/{project}/locations/{location}/agents/{agent_id})
            return f"projects/{self.project_id}/locations/{self.location}/agents/{agent_name}"
            
        except Exception as e:
            logger.error(f"Failed to register agent {agent_name}: {e}")
            raise
    
    def discover_agent(self, agent_name: str) -> Optional[Dict[str, Any]]:
        """
        Discover an agent from Vertex AI Agent Registry
        
        Args:
            agent_name: Name of the agent to discover
        
        Returns:
            Agent information including endpoint, or None if not found
        """
        try:
            # In production, this would query Vertex AI Agent Registry
            # For now, return None (should be implemented with actual registry lookup)
            logger.warning(f"Agent discovery for {agent_name} not fully implemented")
            return None
            
        except Exception as e:
            logger.error(f"Failed to discover agent {agent_name}: {e}")
            return None
    
    def list_agents(self, filter_capabilities: Optional[List[str]] = None) -> List[Dict[str, Any]]:
        """
        List all registered agents, optionally filtered by capabilities
        
        Args:
            filter_capabilities: Optional list of capabilities to filter by
        
        Returns:
            List of agent information
        """
        try:
            # In production, this would query Vertex AI Agent Registry
            # For now, return empty list
            logger.warning("Agent listing not fully implemented")
            return []
            
        except Exception as e:
            logger.error(f"Failed to list agents: {e}")
            return []
    
    def update_agent_status(self, agent_name: str, status: str) -> bool:
        """
        Update agent status in registry
        
        Args:
            agent_name: Name of the agent
            status: New status (ACTIVE, INACTIVE, etc.)
        
        Returns:
            True if successful
        """
        try:
            logger.info(f"Updated agent {agent_name} status to {status}")
            return True
        except Exception as e:
            logger.error(f"Failed to update agent status: {e}")
            return False





